requirements.txt
djangorestframework=3.14.0
django=4.2.5
markdown=3.4.4
django-filter=23.0
pynamodb=5.5.0
boto3=1.28.45
coverage==7.3.1